feat(networkService): Separated DTO into several files
