+++ Recipely/Recipely/Services/NetworkService/DTO/DishTDO.swift
@@ -7,19 +7,6 @@
 
 import Foundation
 
-
-/// Структура для декодирования JSON-ответа от сервера.
-struct Response: Codable {
-    /// Массив объектов типа `Hit`, содержащих информацию о блюдах.
-    let hits: [Hit]
-}
-
-/// Структура, представляющая отдельный элемент массива `hits` в JSON-ответе.
-struct Hit: Codable {
-    /// Объект типа `DishDTO`, содержащий информацию о конкретном блюде.
-    let recipe: DishDTO
-}
-
 /// Структура, представляющая информацию о конкретном блюде.
 struct DishDTO: Codable {
     /// URI блюда.
@@ -37,10 +24,3 @@ struct DishDTO: Codable {
     /// Общие питательные вещества блюда, представленные в виде словаря, где ключ - название питательного вещества, значение - объект типа `Total`.
     let totalNutrients: [String: Total]
 }
-
-/// Структура, представляющая общее количество питательного вещества.
-struct Total: Codable {
-    /// Количество питательного вещества.
-    let quantity: Double
-}
-
