+++ Recipely/Recipely/Services/NetworkService/DTO/Total.swift
@@ -1,0 +1,14 @@
+//
+//  Total.swift
+//  Recipely
+//
+//  Created by Tixon Markin on 13.03.2024.
+//
+
+import Foundation
+
+/// Структура, представляющая общее количество питательного вещества.
+struct Total: Codable {
+    /// Количество питательного вещества.
+    let quantity: Double
+}
